extcon: max77693: Bug fix for irq handler, call handler depends on ADC.
Since it has called charger's or adc's handler depends on irq type, sometimes
it failed to call proper handler. For example, during the driver's probing,
it set ADC debouce time to 25ms by default, then it occurs HW interrupt
which is flagged as CHGTYP. In previous irq handler it called charger handler
because CHGTYP irq is classified as a charger's. But it even happens when
other accessories had been connected. So it is better to call seperate handler
depends on ADC's value.

If ADC is 0x1F, it means that charger is possibly connected so charger's
handler might be needed. Otherwise different handler would be called such as
JIG's or DOCK's and so on.

Signed-off-by: Jonghwa Lee <jonghwa3.lee@samsung.com>
Signed-off-by: Myungjoo Ham <myungjoo.ham@samsung.com>
[Artur: fixed to apply on top of latest driver version]
Signed-off-by: Artur Weber <aweber.kernel@gmail.com>