power_supply: max77693: Listen for cable events and enable charging
This patch adds a listener for extcon cable events and enables
charging if an USB cable is connected. It recognizes SDP and DCP cable
types and treats them the same (same input current and fast charge
current). The maximum input current is set before the charger is
enabled and before the charger gets disabled, the maximum input
current is set to zero. The listener is inspired by the listener
implementation that was used for the AXP288 Charger driver.

Signed-off-by: Wolfgang Wiedmeyer <wolfgit@wiedmeyer.de>
[GNUtoo@cyberdimension.org: small fixes]
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
[Artur: dropped CURRENT_NOW property, dropped current change in enable func]
Signed-off-by: Artur Weber <aweber.kernel@gmail.com>