ASoC: dt-bindings: samsung,midas-audio: Add GPIO-based headset jack detection
Some Samsung devices that share the midas-audio driver use a GPIO-based
approach to headset jack detection, as opposed to using the built-in
jack detection provided by the wm8994 driver. This setup uses two GPIOs
(one for jack detection and another for key detection) and an ADC
channel for determining the jack type or button pressed.

Add DT configuration values that allow for describing these setups.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
Changes in v3:
- Re-added pipe (|) to description of samsung,headset-button-threshold-
  microvolt to prevent syntax error (otherwise the "in order: ..." part
  is interpreted as a list)
Changes in v2:
- Added vendor prefix to threshold properties
- Dropped pipe (|) character from description: field