power: supply: max77693: Set fast charge current at initialization time
Set up the fast charge current limit register to either a value provided
in the DTS (maxim,fast-charge-current-microamps property), or a safe
default value of 500mA.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
It's worth noting that the downstream driver[1] actually sets the
default value to 0; however, it appears to be set back to 500mA later on.
It does this, as it provides this data in the CURRENT_NOW variable,
which isn't really correct at this doesn't guarantee that the current
will be 500mA minimum, but maximum.

[1] https://github.com/gr8nole/android_kernel_samsung_smdk4x12/blob/lineage-14.1/drivers/battery/max77693_charger_bat.c#L850-L858