power: supply: max77693: Expose INPUT_CURRENT_LIMIT property
Add a function to get the current limit by checking the fast charge
current limit and the CHGIN input limit.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
A question - is my understanding of this correct? Unfortunately
the downstream kernel has pretty much no documentation about any of
this, and I do not have access to the datasheet to verify.

To be more precise - what is the difference between fast charge
current limit and CHGIN input limit? Which one comes first?