power: supply: max77693: Set up charge/input current according to cable type
This behavior was observed on a downstream kernel - for chargers, the
current would be set to a fast charge current value, and it would be
bumped down for all other cable types.

If we leave only the fast charge current value applied, peripheral mode
stops working. If we stick to 500mA, charging becomes too slow. So, set
the charge input current limits accordingly to the cable type.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>