drm/panel: samsung-s6d7aa0: use 58Hz refresh rate for lsl080al02
The downstream kernel for this device used 58Hz as the refresh rate[1],
but I bumped it up to 60Hz as seemed to work fine on one of my tablets.
However, on another unit with a more worn-down screen, it causes weird
display corruption.

Change the refresh rate in the clock calculation to 58Hz to match
downstream and fix display corruption on worn-down panels.

[1] https://github.com/gr8nole/android_kernel_samsung_smdk4x12/blob/lineage-14.1/arch/arm/mach-exynos/board-lcd-mipi.c#L585

Fixes: 6810bb390282 ("drm/panel: Add Samsung S6D7AA0 panel controller driver")
Signed-off-by: Artur Weber <aweber.kernel@gmail.com>