[RFC] power: supply: max77693: Toggle charging/OTG based on extcon status
This patchset does the following:

- Add CURRENT_MAX and INPUT_CURRENT_MAX power supply properties to
  expose the "fast charge current" (maximum current from charger to
  battery) and "CHGIN input current limit" (maximum current from
  external supply to charger).

- Add functions for toggling charging and OTG modes.

- Add an extcon-based handler that enables charging or OTG depending
  on the cable type plugged in. The extcon device to use for cable
  detection can be specified in the device tree, and is entirely
  optional.

The extcon listener implementation is inspired by the rt5033 charger
driver (commit 8242336dc8a8 ("power: supply: rt5033_charger: Add cable
detection and USB OTG supply")).

To: Krzysztof Kozlowski <krzk@kernel.org>
Cc: Sebastian Reichel <sre@kernel.org>
Cc: Rob Herring <robh@kernel.org>
Cc: Conor Dooley <conor+dt@kernel.org>
To: Chanwoo Choi <cw00.choi@samsung.com>
Cc: Lee Jones <lee@kernel.org>
Cc: Krzysztof Kozlowski <krzk+dt@kernel.org>
Cc: Alim Akhtar <alim.akhtar@samsung.com>
Cc: linux-pm@vger.kernel.org
Cc: devicetree@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-samsung-soc@vger.kernel.org
Cc: ~postmarketos/upstreaming@lists.sr.ht
Cc: Henrik Grimler <henrik@grimler.se>
Cc: Wolfgang Wiedmeyer <wolfgit@wiedmeyer.de>
Cc: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
Quick recap - MAX77693 charger has 2 "current limit" registers:

- Input current limit (cable -> charger), managed by CHARGER regulator
  of the MAX77693 MFD
- CC/Fast charge current (charger -> battery)

I initially agreed to the suggestion of rolling up both of these registers
into the CHARGER regulator, but I've had some time to think about it since
and I'm not sure that I'm entirely happy with that.

For one, this is kind of unsafe. The current behavior of the CHARGER
regulator is to only set the **input** current limit; the actual current
going to the battery is still capped by the fast charge current register.
If you set the input current too high, it's cushioned by the CC limit.
If we were to change the CHARGER regulator to modify both values,
a bad value that didn't cause issues before would instead cause potential
damage to the battery. (Though, I suppose it was expected behavior anyways
that both registers would be set, judging by the TODO in the regulator
driver... maybe the kernel shouldn't be concerned about user error.)

Either way, it would be safer to let userspace only modify the input
current limit, and have the fast charge current limit set once at init
based on a hardcoded value in the DT. I think they should be considered
two separate values (even the power supply API seems to agree, given
that CURRENT_MAX and INPUT_CURRENT_LIMIT are two separate properties).

(Also, using a regulator like this is opaque and implies that it can be
replaced by some external regulator, which isn't really the case. This
patch explicitly used the regulator to manipulate the charger registers,
which is better done directly. It also led to some silly issues like
the hardware-enabled state becoming out-of-sync with the regulator
status, etc.)

My proposal is to thus keep the CHARGER regulator for the old userspace
but deprecate it in favor of managing the charging parameters through
power supply properties, as is standard practice in other drivers.
(Still RFC, comments/opinions appreciated!)
---
Changes in v2:
- Stop using the CHARGER regulator in the power supply driver in favor
  of modifying registers directly
- Link to v1: https://lore.kernel.org/r/20240530-max77693-charger-extcon-v1-0-dc2a9e5bdf30@gmail.com

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240525-max77693-charger-extcon-9ebb7bad83ce",
    "prefixes": [],
    "history": {
      "v1": [
        "20240530-max77693-charger-extcon-v1-0-dc2a9e5bdf30@gmail.com"
      ]
    }
  }
}