phy: samsung-usb2: Add extcon support for handling mode switch
The Exynos 4x12 and 5250/5420 have a "mode switch" register which is
used to toggle between "device" and "host" modes. This seems to be
necessary to get the "hsotg" or host mode (EHCI/OHCI) to work.

This switch is toggled in the driver while powering on/off the PHYs.
Add the ability to override this toggle by monitoring the status of
an extcon device representing a USB connector. If a cable of type
EXTCON_USB_HOST is plugged in, it switches to host mode, otherwise it
switches to device mode. (This is because the hsotg controller is more
sensitive to the mode being set to "host" than the EHCI controller
is to the mode being set to "device".)

This mechanism is completely optional, so it shouldn't break existing
devices.

---

This fixes an issue on the Exynos 4x12 where toggling both the hsotg
and ehci nodes at the same time would break peripheral mode. This SoC
has *two* USB controllers: a peripheral-only hsotg/dwc2 controller,
and an ehci/ohci controller for host mode. To get USB OTG to work,
both of these have to be enabled; however, the hsotg/dwc2 controller
would complain whe

The issue looked something like this:

- hsotg/dwc2 gets probed
  -> hsotg depends on device PHY
  -> device PHY is powered on, mode switch set to device
- ehci/ohci gets probed
  -> ehci depends on host PHY
  -> host PHY is powered on, mode switch set to host
- dwc2 driver performs delayed gadget init once the USB gadget driver
  is loaded
  -> since the mode switch has been set to host by ehci init, this
     fails with a soft reset timeout - hsotg controller is down!

As far as I can tell, the mode switch toggles the entire controller
on/off, not just the PHYs (or at least it does some power management
that causes the dwc2 controller to not respond).

This patchset works around the issue by allowing the use of an
extcon device to override the automatic mode switch.

This could theoretically be fixed by toggling the host PHY only,
at least on the Exynos 4x12. But that would require modifying the
EHCI/OHCI driver to handle every PHY separately (I think HCD core
does bulk operations on all PHYs, EHCI driver also duplicates this
for some legacy reasons).

This is required to get OTG working on the Galaxy S3 and other
Exynos 4x12 devices in mainline. It's been tested on the Samsung
Galaxy Tab 3 8.0, an Exynos 4212 device.

To: Sylwester Nawrocki <s.nawrocki@samsung.com>
To: Vinod Koul <vkoul@kernel.org>
To: Kishon Vijay Abraham I <kishon@kernel.org>
To: Rob Herring <robh@kernel.org>
To: Krzysztof Kozlowski <krzk+dt@kernel.org>
To: Conor Dooley <conor+dt@kernel.org>
To: Marek Szyprowski <m.szyprowski@samsung.com>
To: Alim Akhtar <alim.akhtar@samsung.com>
Cc: Krzysztof Kozlowski <krzk@kernel.org>
Cc: linux-kernel@vger.kernel.org
Cc: linux-phy@lists.infradead.org
Cc: devicetree@vger.kernel.org
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-samsung-soc@vger.kernel.org
Cc; ~postmarketos/upstreaming@lists.sr.ht
Signed-off-by: Artur Weber <aweber.kernel@gmail.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240619-exynos-phy-extcon-eec2fd13eee0",
    "prefixes": []
  }
}