power: supply: max77693: Allow for setting input current limit property
Allow for setting the input current limit of the MAX77963 charger by
setting the INPUT_CURRENT_LIMIT property of the power supply. This
allows for fine-tuning the input current limit from userspace, but
keeps the actual current limit going to the battery intact, which
is safer than setting both values.

Old userspace used to do this with the CHARGER regulator of the main
MAX77693 device. This supersedes it with a solution that's more in-line
with other power supply drivers.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>
---
Changes in v2:
- Added this commit