phy: samsung-usb2: Add common interface for USB mode switch
The Exynos 4x12 and 5250/5420 have a "mode switch" register which is
used to toggle between "device" and "host" modes. This switch is
necessary to get the "hsotg" (peripheral) or host mode (EHCI/OHCI)
controllers to work.

Move the logic for setting these modes to separate "set mode switch"
functions rather than modifying the regmap directly, and keep those
functions in the SoC-specific PHY configs for easier access.

This is done in preparation for an extcon-based mode switch mechanism
which will use these functions.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>